# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EPlusActivities.API",
    "version": "v1"
  },
  "paths": {
...
    "/choujiang/api/ActivityUser": {
      "get": {
        "tags": [
          "ActivityUser"
        ],
        "summary": "获取活动和用户的绑定关系",
        "parameters": [
          {
            "name": "UserId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "ActivityId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityUserDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityUserDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityUserDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ActivityUser"
        ],
        "summary": "绑定单个活动和用户，该 API 必须在签到和抽奖 API 之前被调用，否则无法判断用户参与的是哪个活动。",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindActivityAndUserCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BindActivityAndUserCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BindActivityAndUserCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/choujiang/api/ActivityUser/activities": {
      "get": {
        "tags": [
          "ActivityUser"
        ],
        "summary": "获取某个用户正在参与的活动",
        "parameters": [
          {
            "name": "UserId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "StartTime",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            }
          },
          {
            "name": "EndTime",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            }
          },
          {
            "name": "AvailableChannel",
            "in": "query",
            "description": "用户访问的渠道，字符串不区分大小写。\r\n取值范围：Ngs, NgsPlaza, Alldays",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ChannelCode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityUserDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityUserDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityUserDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/choujiang/api/ActivityUser/bindingAvailableActivities": {
      "post": {
        "tags": [
          "ActivityUser"
        ],
        "summary": "为某个用户绑定所有可参加的活动",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindAvailableActivitiesCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BindAvailableActivitiesCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BindAvailableActivitiesCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityUserDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityUserDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityUserDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/choujiang/api/ActivityUser/redeeming": {
      "patch": {
        "tags": [
          "ActivityUser"
        ],
        "summary": "用户积分兑换抽奖次数",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityUserForRedeemDrawsResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityUserForRedeemDrawsResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityUserForRedeemDrawsResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivityUserDto": {
        "type": "object",
        "properties": {
          "usedDraws": {
            "type": "integer",
            "description": "全活动周期已抽奖次数",
            "format": "int32"
          },
          "todayUsedDraws": {
            "type": "integer",
            "description": "当天已抽奖次数",
            "format": "int32"
          },
          "remainingDraws": {
            "type": "integer",
            "description": "剩余抽奖次数，null 表示无限",
            "format": "int32"
          },
          "attendanceDays": {
            "type": "integer",
            "description": "签到天数，非签到活动中用 null 表示",
            "format": "int32",
            "nullable": true
          },
          "sequentialAttendanceDays": {
            "type": "integer",
            "description": "连续签到天数，非签到活动中用 null 表示",
            "format": "int32",
            "nullable": true
          },
          "requiredCreditForRedeeming": {
            "type": "integer",
            "description": "兑换一次抽奖所需积分，null 表示非抽奖活动",
            "format": "int32",
            "nullable": true
          },
          "lastAttendanceDate": {
            "type": "string",
            "description": "上次签到日期",
            "format": "date-time",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "activityId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BindActivityAndUserCommand": {
        "required": [
          "activityId",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "activityId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "ChannelCode": {
        "enum": [
          "Ngs",
          "NgsPlaza",
          "Alldays",
          "Test"
        ],
        "type": "string"
      },
      "BindAvailableActivitiesCommand": {
        "required": [
          "channelCode",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "channelCode": {
            "$ref": "#/components/schemas/ChannelCode"
          }
        },
        "additionalProperties": false
      },
      "RedeemCommand": {
        "required": [
          "activityId",
          "channel",
          "userId"
        ],
        "type": "object",
        "properties": {
          "activityId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "channel": {
            "$ref": "#/components/schemas/ChannelCode"
          },
          "count": {
            "type": "integer",
            "description": "兑换几次",
            "format": "int32"
          },
          "reason": {
            "type": "string",
            "description": "原因",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ActivityUserForRedeemDrawsResponseDto": {
        "required": [
          "activityId",
          "userId"
        ],
        "type": "object",
        "properties": {
          "activityId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "remainingDraws": {
            "type": "integer",
            "description": "剩余抽奖次数",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}